dustbin/requirements.txt
opencv-python
Pillow
google-generativeai
python-dotenv
